Add deepnote/ir-with-libs image
Based on the list of packages in https://community.deepnote.com/c/showcase/plotly-in-r-project#comment_wrapper_949599 plus some I got from another used on Crisp (from the same team as the community post author).

The ideal way to implement this would be to have this installed in a mountable directory and mount it to the image same as what we do with Python packages, but that would be more difficult to set up (we would need to validate it is compatible with different distros etc). This was a (almost) quick win.

The plan is to let people in community know about this image so that they can use it as Custom Docker Image. Optionally we can also offer it in the dropdown of the Deepnote images (and add it to the prepull at the same time).